cleanup routing API and examples on transits

+++ ortools/constraint_solver/routing_types.h
@@ -37,8 +37,8 @@ DEFINE_INT_TYPE(RoutingDimensionIndex, int);
 DEFINE_INT_TYPE(RoutingDisjunctionIndex, int);
 DEFINE_INT_TYPE(RoutingVehicleClassIndex, int);
 
-typedef std::function<int64(int)> RoutingTransitCallback1;
-typedef std::function<int64(int, int)> RoutingTransitCallback2;
+typedef std::function<int64(int64)> RoutingTransitCallback1;
+typedef std::function<int64(int64, int64)> RoutingTransitCallback2;
 // NOTE(user): keep the "> >" for SWIG.
 typedef std::pair<std::vector<int64>, std::vector<int64> > RoutingIndexPair;
 typedef std::vector<RoutingIndexPair> RoutingIndexPairs;
